Install: Add list access check
Add a validation step after creating the service principal to confirm that
it can ebe used to access the users, groups, and SP list APIs. From
testing and investigation, this endpoint should always be available
provided the SP can authenticate. This step allows us to detect if this
assumption has somehow changed early as it will break access control later.

- Update the docs url in the README as it currently points to a page that
  returns a 404
- add a pip install for the dasl-client to address `No module named
  'dasl_client'` when running the notebook.
- Update `securable_type` values to be a securable type rather than the
  type name. Providing just the name causes a `missing values` error whe
  running.